[WIP] Default Branches: Permit repos to have a default branch
Currently, the way of determining which branch should be used, should
the branch CLI option not be provided, is to use the currently checked
out branch. This makes it difficult to specify a policy for a suite, or
even have a different branch for each suite.

Add the beginnings of an approach to permit default branches to be
specified by providing an option in the suite definitions which can be
overridden by the package specific options.