Performance: Run lengthy tasks in threads
Utilise a thread pool for running tasks which outsource computation to
the `reprepro` CLI tool which waits while the shell command returns and
the output is parsed. Apply this fully to the `suite-diff` command, and
as a side-effect for any command which at some point utilises the
`Repo::list_packages()` function, to dramatically reduce the time taken
to run.

Test results reduced the `suite-diff` time from ~10 minutes to ~100
seconds.