Repo: Add mutex lock for reprepro calls
When querying packages for their versions, a series of threads are
created to speed up the process and each calls the appropriate
`reprepro` command to query this information. This would occasionally
fail as only one command can run at a time so the original strategy was
to retry a few times. Instead of this, create a mutex for the command
and acquire that before attempting to completely avoid problems.