Build: Pass environment arguments to correct command
In 498dfd26fde1bb53d33073e831eb87eb417e25fa, an argument was passed to
the `sudo` command during build to preserve the environment variables.
This was accidentally passed to the `chroot` command that followed, not
the `sudo` command, thus causing the build to explode. Resolve this by
passing the argument to the correct command.